+++ presentation/src/main/AndroidManifest.xml
@@ -16,14 +16,14 @@
             android:theme="@style/AppTheme"
             tools:ignore="GoogleAppIndexingWarning">
 
-        <activity android:name=".splash.SplashActivity">
+        <activity android:name=".views.splash.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
 
-        <activity android:name=".main.MainActivity" />
+        <activity android:name=".views.main.MainActivity" />
 
     </application>
