+++ presentation/src/main/kotlin/com/mickeycj/restcountries/views/main/countries/CountriesFragment.kt
@@ -1,19 +1,20 @@
-package com.mickeycj.restcountries.main.countrydetails
+package com.mickeycj.restcountries.views.main.countries
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+
 import com.mickeycj.restcountries.R
 
 /**
- * Fragment for Country Details screen.
+ * Fragment for Countries screen.
  */
-class CountryDetailsFragment : Fragment() {
+class CountriesFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.fragment_country_details, container, false)
+    ): View? = inflater.inflate(R.layout.fragment_countries, container, false)
 }
