+++ presentation/src/main/kotlin/com/mickeycj/restcountries/views/main/countries/CountriesViewModel.kt
@@ -1,4 +1,4 @@
-package com.mickeycj.restcountries.main.countries
+package com.mickeycj.restcountries.views.main.countries
 
 import io.reactivex.Scheduler
 
@@ -7,9 +7,9 @@ import androidx.lifecycle.MutableLiveData
 import com.mickeycj.domain.models.Country
 import com.mickeycj.domain.usecases.GetCountriesUseCase
 
-import com.mickeycj.restcountries.common.StateMapper
-import com.mickeycj.restcountries.common.ViewModel
-import com.mickeycj.restcountries.states.CountriesState
+import com.mickeycj.restcountries.common.base.StateMapper
+import com.mickeycj.restcountries.common.base.ViewModel
+import com.mickeycj.restcountries.common.states.CountriesState
 
 /**
  * View Model for Countries screen.
