+++ presentation/src/main/kotlin/com/mickeycj/restcountries/views/main/countrydetails/CountryDetailsFragment.kt
@@ -1,19 +1,20 @@
-package com.mickeycj.restcountries.main.countries
+package com.mickeycj.restcountries.views.main.countrydetails
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+
 import com.mickeycj.restcountries.R
 
 /**
- * Fragment for Countries screen.
+ * Fragment for Country Details screen.
  */
-class CountriesFragment : Fragment() {
+class CountryDetailsFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.fragment_countries, container, false)
+    ): View? = inflater.inflate(R.layout.fragment_country_details, container, false)
 }
