+++ presentation/src/main/kotlin/com/mickeycj/restcountries/views/main/countrydetails/CountryDetailsViewModel.kt
@@ -1,4 +1,4 @@
-package com.mickeycj.restcountries.main.countrydetails
+package com.mickeycj.restcountries.views.main.countrydetails
 
 import io.reactivex.Scheduler
 
@@ -7,9 +7,9 @@ import androidx.lifecycle.MutableLiveData
 import com.mickeycj.domain.models.Country
 import com.mickeycj.domain.usecases.GetCountryDetailsUseCase
 
-import com.mickeycj.restcountries.common.StateMapper
-import com.mickeycj.restcountries.common.ViewModel
-import com.mickeycj.restcountries.states.CountryDetailsState
+import com.mickeycj.restcountries.common.base.StateMapper
+import com.mickeycj.restcountries.common.base.ViewModel
+import com.mickeycj.restcountries.common.states.CountryDetailsState
 
 /**
  * View Model for Country Details screen.
