+++ presentation/src/main/res/layout/fragment_countries.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".main.countries.CountriesFragment">
+        tools:context=".views.main.countries.CountriesFragment">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
             android:layout_width="match_parent"
